[core] -Cleanup: api version check printf warning on 64bits
git-svn-id: svn://repo.hu/pcb-rnd/trunk@16271 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/plugins.h
@@ -37,7 +37,7 @@ extern unsigned long pcb_api_ver;
 #define PCB_API_CHK_VER \
 do { \
 	if (!PCB_API_VER_MATCH) {\
-		fprintf(stderr, "pcb-rnd API version incompatibility: " __FILE__ "=%lu core=%lu\n(not loading this plugin)\n", PCB_API_VER, pcb_api_ver); \
+		fprintf(stderr, "pcb-rnd API version incompatibility: " __FILE__ "=%lu core=%lu\n(not loading this plugin)\n", (unsigned long)PCB_API_VER, pcb_api_ver); \
 		return 1; \
 	} \
 } while(0)
